mobile navbar redone using seperate HTML elements
Before this, JS and CSS Media Queries were used for the same elements to
re-organize them for the mobile navbar menu. This caused problems for when
tabbing through elements. Now, seperate elements are
used (set to display on media Query) and the tabbing sequence is more
appropriate.

Also fixed: navbar not being sticky (was due to hidden overflow on
parent).